libs/common_logging/pyproject.toml
[tool.poetry]
name = "common_logging"
version = "0.1.0"
description = "Shared logging utilities for home-agent-suite monorepo"
authors = ["davidasnider <david@davidsnider.org>"]

[tool.poetry.dependencies]
python = ">=3.9"
google-cloud-logging = "^3.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
